Amazing hacks to support shell-wrappers for testing. Details below.
Login as an administrator (e.g., arjun@cs.brown.edu) and choose "Create
New Assignment". There is a new type of assignment called "shell".
Shell solutions are handled by a shell script.

A run involves a test suite from a user A and a solution from a user B.
When a shell-assignment is run, the test suite is placed in a temporary
directory, in a file called "thetestsuite.ml". *All* the latest
solutions by user B are placed in the same directory, in files named
after the assignment.

Tip 1: Do not use "thetestsuite.ml" as an assignment name.

Tip 2: Consider using an extension such as .ml in the assignment name
       itself.

After the directory is prepared, the script ../scripts/camlharness is
called with two command-line arguments: the name of the temporary
directory and the assignment name. Given the assignment name and the
well-known filenames in the directory, the script should be able to run
the test.  If the script terminates with exit code 0, then the test
passes. Otherwise, the test fails.

Tip 3: The current camlharness script prints its command line arguments
       and the contents of files in the test directory. This should give
       you a feel for the testing environment.

"Gold" Solutions
----------------

"Gold" solutions are perfect, TA-written solutions used to determine if
student-submitted test suites are correct.  If a student-submitted test
fails the gold solution, it is automatically rejected from the pool of
global tests.  When you create a shell test, instead of entering a
"gold" solution, you enter the username of a designated TA who is
charged with submitting perfect solutions. (Good luck with that.)

TestFest will accept solutions by this designated TA account without any
testing.